+++ lib/err/test/test__exit_codes.ml
@@ -52,8 +52,19 @@ let%expect_test "exit" =
 
 let%expect_test "exit without handler" =
   require_does_raise [%here] (fun () -> Err.exit Err.Exit_code.ok);
-  [%expect {| ((Exit 0)) |}];
+  [%expect {| (exit_code 0) |}];
   require_does_raise [%here] (fun () -> Err.exit Err.Exit_code.some_error);
-  [%expect {| ((Exit 123)) |}];
+  [%expect {| (exit_code 123) |}];
+  ()
+;;
+
+let%expect_test "exit_code in sexp" =
+  let e = Err.create [ Pp.text "Hello Exit Code" ] in
+  print_s [%sexp (e : Err.t)];
+  [%expect {| "Hello Exit Code" |}];
+  require_does_raise [%here] (fun () -> raise (Err.E e));
+  [%expect {| "Hello Exit Code" |}];
+  print_s [%sexp (e : Err.With_exit_code.t)];
+  [%expect {| ("Hello Exit Code" (exit_code 123)) |}];
   ()
 ;;